+++ lib/react_on_rails/helper.rb
@@ -17,6 +17,9 @@ module ReactOnRails
     include ReactOnRails::Utils::Required
 
     COMPONENT_HTML_KEY = "componentHtml"
+    IMMEDIATE_HYDRATION_PRO_WARNING = "[REACT ON RAILS] The 'immediate_hydration' feature requires a " \
+                                      "React on Rails Pro license. " \
+                                      "Please visit https://shakacode.com/react-on-rails-pro to learn more."
 
     # react_component_name: can be a React function or class component or a "Render-Function".
     # "Render-Functions" differ from a React function in that they take two parameters, the
@@ -58,7 +61,7 @@ module ReactOnRails
       server_rendered_html = internal_result[:result]["html"]
       console_script = internal_result[:result]["consoleReplayScript"]
       render_options = internal_result[:render_options]
-      badge = pro_warning_badge_if_needed(render_options.force_load)
+      badge = pro_warning_badge_if_needed(internal_result[:immediate_hydration_requested])
 
       case server_rendered_html
       when String
@@ -128,7 +131,7 @@ module ReactOnRails
       # stream_react_component doesn't have the prerender option
       # Because setting prerender to false is equivalent to calling react_component with prerender: false
       options[:prerender] = true
-      options = options.merge(force_load: true) unless options.key?(:force_load)
+      options = options.merge(immediate_hydration: true) unless options.key?(:immediate_hydration)
       run_stream_inside_fiber do
         internal_stream_react_component(component_name, options)
       end
@@ -210,11 +213,12 @@ module ReactOnRails
     #
     def react_component_hash(component_name, options = {})
       options[:prerender] = true
+
       internal_result = internal_react_component(component_name, options)
       server_rendered_html = internal_result[:result]["html"]
       console_script = internal_result[:result]["consoleReplayScript"]
       render_options = internal_result[:render_options]
-      badge = pro_warning_badge_if_needed(render_options.force_load)
+      badge = pro_warning_badge_if_needed(internal_result[:immediate_hydration_requested])
 
       if server_rendered_html.is_a?(String) && internal_result[:result]["hasErrors"]
         server_rendered_html = { COMPONENT_HTML_KEY => internal_result[:result]["html"] }
@@ -252,15 +256,16 @@ module ReactOnRails
     # props: Ruby Hash or JSON string which contains the properties to pass to the redux store.
     # Options
     #    defer: false -- pass as true if you wish to render this below your component.
-    #    force_load: false -- pass as true if you wish to hydrate this store immediately instead of
-    #                        waiting for the page to load.
-    def redux_store(store_name, props: {}, defer: false, force_load: nil)
-      force_load = ReactOnRails.configuration.force_load if force_load.nil?
-      badge = pro_warning_badge_if_needed(force_load)
+    #    immediate_hydration: false -- React on Rails Pro (licensed) feature. Pass as true if you wish to
+    #                        hydrate this store immediately instead of waiting for the page to load.
+    def redux_store(store_name, props: {}, defer: false, immediate_hydration: nil)
+      immediate_hydration = ReactOnRails.configuration.immediate_hydration if immediate_hydration.nil?
+      badge = pro_warning_badge_if_needed(immediate_hydration)
+      immediate_hydration = false unless support_pro_features?
 
       redux_store_data = { store_name: store_name,
                            props: props,
-                           force_load: force_load }
+                           immediate_hydration: immediate_hydration }
       if defer
         registered_stores_defer_render << redux_store_data
         "YOU SHOULD NOT SEE THIS ON YOUR VIEW -- Uses as a code block, like <% redux_store %> " \
@@ -447,16 +452,20 @@ module ReactOnRails
 
     # rubocop:enable Metrics/AbcSize, Metrics/CyclomaticComplexity
 
-    def pro_warning_badge_if_needed(force_load)
-      return "".html_safe unless force_load
-      return "".html_safe if ReactOnRails::Utils.react_on_rails_pro_licence_valid?
+    # Checks if React on Rails Pro features are available
+    # @return [Boolean] true if Pro license is valid, false otherwise
+    def support_pro_features?
+      ReactOnRails::Utils.react_on_rails_pro_licence_valid?
+    end
 
-      warning_message = "[REACT ON RAILS] The 'force_load' feature requires a React on Rails Pro license. " \
-                        "Please visit https://shakacode.com/react-on-rails-pro to learn more."
-      puts warning_message
-      Rails.logger.warn warning_message
+    def pro_warning_badge_if_needed(immediate_hydration)
+      return "".html_safe unless immediate_hydration
+      return "".html_safe if support_pro_features?
 
-      tooltip_text = "The 'force_load' feature requires a React on Rails Pro license. Click to learn more."
+      puts IMMEDIATE_HYDRATION_PRO_WARNING
+      Rails.logger.warn IMMEDIATE_HYDRATION_PRO_WARNING
+
+      tooltip_text = "The 'immediate_hydration' feature requires a React on Rails Pro license. Click to learn more."
 
       badge_html = <<~HTML
         <a href="https://shakacode.com/react-on-rails-pro" target="_blank" rel="noopener noreferrer" title="#{tooltip_text}">
@@ -666,6 +675,9 @@ module ReactOnRails
       # server has already rendered the HTML.
 
       render_options = create_render_options(react_component_name, options)
+      # Capture the originally requested value so we can show a badge while still disabling the feature.
+      immediate_hydration_requested = render_options.immediate_hydration
+      render_options.set_option(:immediate_hydration, false) unless support_pro_features?
 
       # Setup the page_loaded_js, which is the same regardless of prerendering or not!
       # The reason is that React is smart about not doing extra work if the server rendering did its job.
@@ -678,9 +690,10 @@ module ReactOnRails
                                                 "data-trace" => (render_options.trace ? true : nil),
                                                 "data-dom-id" => render_options.dom_id,
                                                 "data-store-dependencies" => render_options.store_dependencies&.to_json,
-                                                "data-force-load" => (render_options.force_load ? true : nil))
+                                                "data-immediate-hydration" =>
+                                                  (render_options.immediate_hydration ? true : nil))
 
-      if render_options.force_load
+      if render_options.immediate_hydration
         component_specification_tag.concat(
           content_tag(:script, %(
 typeof ReactOnRails === 'object' && ReactOnRails.reactOnRailsComponentLoaded('#{render_options.dom_id}');
@@ -695,7 +708,8 @@ typeof ReactOnRails === 'object' && ReactOnRails.reactOnRailsComponentLoaded('#{
       {
         render_options: render_options,
         tag: component_specification_tag,
-        result: result
+        result: result,
+        immediate_hydration_requested: immediate_hydration_requested
       }
     end
 
@@ -704,9 +718,10 @@ typeof ReactOnRails === 'object' && ReactOnRails.reactOnRailsComponentLoaded('#{
                                          json_safe_and_pretty(redux_store_data[:props]).html_safe,
                                          type: "application/json",
                                          "data-js-react-on-rails-store" => redux_store_data[:store_name].html_safe,
-                                         "data-force-load" => (redux_store_data[:force_load] ? true : nil))
+                                         "data-immediate-hydration" =>
+                                           (redux_store_data[:immediate_hydration] ? true : nil))
 
-      if redux_store_data[:force_load]
+      if redux_store_data[:immediate_hydration]
         store_hydration_data.concat(
           content_tag(:script, <<~JS.strip_heredoc.html_safe
             typeof ReactOnRails === 'object' && ReactOnRails.reactOnRailsStoreLoaded('#{redux_store_data[:store_name]}');